feat: SetZero test

+++ allocator/include/mpool/vmm_allocator.h
@@ -29,7 +29,7 @@ struct VMMAllocatorConfig {
   std::string belong_name;
   size_t small_block_nbytes;
   size_t align_nbytes;
-  int belong_id;
+  long belong_id;
 };
 
 class VMMAllocator : public IAllocator {
